+++ src/simulation/basic/discrete.rs
@@ -9,7 +9,7 @@ pub trait DiscreteProcess: Send + Sync + Sized {
     /// # Arguments
     ///
     /// * `num_step` - The number of steps of the simulation.
-    fn simulate(&self, num_step: u64) -> XResult<DiscretePair>;
+    fn simulate(&self, num_step: usize) -> XResult<DiscretePair>;
 
     /// Get the mean of the discrete process
     ///
@@ -17,7 +17,7 @@ pub trait DiscreteProcess: Send + Sync + Sized {
     ///
     /// * `num_step` - The number of steps of the simulation.
     /// * `particles` - The number of particles.
-    fn mean(&self, num_step: u64, particles: u64) -> XResult<f64> {
+    fn mean(&self, num_step: usize, particles: usize) -> XResult<f64> {
         let traj = self.step(num_step)?;
         traj.raw_moment(1, particles, 0.1)
     }
@@ -28,7 +28,7 @@ pub trait DiscreteProcess: Send + Sync + Sized {
     ///
     /// * `num_step` - The number of steps of the simulation.
     /// * `particles` - The number of particles.
-    fn msd(&self, num_step: u64, particles: u64) -> XResult<f64> {
+    fn msd(&self, num_step: usize, particles: usize) -> XResult<f64> {
         let traj = self.step(num_step)?;
         traj.central_moment(2, particles, 0.1)
     }
@@ -40,7 +40,7 @@ pub trait DiscreteProcess: Send + Sync + Sized {
     /// * `num_step` - The number of steps.
     /// * `order` - The order of the moment.
     /// * `particles` - The number of particles.
-    fn raw_moment(&self, num_step: u64, order: i32, particles: u64) -> XResult<f64> {
+    fn raw_moment(&self, num_step: usize, order: i32, particles: usize) -> XResult<f64> {
         let traj = self.step(num_step)?;
         traj.raw_moment(order, particles, 0.1)
     }
@@ -52,7 +52,7 @@ pub trait DiscreteProcess: Send + Sync + Sized {
     /// * `num_step` - The number of steps.
     /// * `order` - The order of the moment.
     /// * `particles` - The number of particles.
-    fn central_moment(&self, num_step: u64, order: i32, particles: u64) -> XResult<f64> {
+    fn central_moment(&self, num_step: usize, order: i32, particles: usize) -> XResult<f64> {
         let traj = self.step(num_step)?;
         traj.central_moment(order, particles, 0.1)
     }
@@ -64,7 +64,7 @@ pub trait DiscreteTrajectoryTrait: DiscreteProcess {
     /// # Arguments
     ///
     /// * `num_step` - The number of steps of the trajectory
-    fn step(&self, num_step: u64) -> XResult<DiscreteTrajectory<Self>> {
+    fn step(&self, num_step: usize) -> XResult<DiscreteTrajectory<Self>> {
         let traj = DiscreteTrajectory::new(self, num_step)?;
         Ok(traj)
     }
@@ -78,12 +78,12 @@ pub struct DiscreteTrajectory<'a, SP: DiscreteProcess> {
     /// The discrete process
     pub(crate) sp: &'a SP,
     /// The number of steps
-    pub(crate) num_step: u64,
+    pub(crate) num_step: usize,
 }
 
 impl<'a, SP: DiscreteProcess> DiscreteTrajectory<'a, SP> {
     /// Create a new `DiscreteTrajetory` with given `DiscreteProcess` and num of steps.
-    pub fn new(sp: &'a SP, num_step: u64) -> XResult<Self> {
+    pub fn new(sp: &'a SP, num_step: usize) -> XResult<Self> {
         if num_step == 0 {
             return Err(SimulationError::InvalidParameters(
                 "num_step must be positive".to_string(),
@@ -99,7 +99,7 @@ impl<'a, SP: DiscreteProcess> DiscreteTrajectory<'a, SP> {
     }
 
     /// Get the number of steps of the trajectory
-    pub fn get_num_step(&self) -> u64 {
+    pub fn get_num_step(&self) -> usize {
         self.num_step
     }
